refactor(reader): dissolve ChapterTimeEstimate into owners
Move PageDwell to a header-only helper, chapter-progress labels onto
CrossPointSettings, and ETA duration formatting into ReaderUtils.
