+++ src/activities/reader/EpubReaderActivity.cpp
@@ -45,7 +45,7 @@
 #include "fontIds.h"
 #include "util/AchievementPopupUtils.h"
 #include "util/BookIdentity.h"
-#include "util/ChapterTimeEstimate.h"
+#include "util/PageDwell.h"
 #include "util/CompletedBookMover.h"
 #include "util/ScreenshotUtil.h"
 
@@ -2032,8 +2032,8 @@ void EpubReaderActivity::renderStatusBar() const {
   if (section->currentPage >= 0) {
     const double wordsPerMs = READING_STATS.getEffectiveWordsPerMs();
     // Skip remaining-words walk when rate is 0 or time is hidden.
-    if (ChapterTimeEstimate::statusBarWantsChapterTime() && wordsPerMs > 0.0 &&
-        ChapterTimeEstimate::formatRemainingFromRate(
+    if (SETTINGS.statusBarWantsChapterTime() && wordsPerMs > 0.0 &&
+        ReaderUtils::formatRemainingFromRate(
             section->estimateRemainingWords(static_cast<uint16_t>(section->currentPage)), wordsPerMs, chapterTimeBuf,
             sizeof(chapterTimeBuf))) {
       chapterTimeEstimate = chapterTimeBuf;
