+++ src/activities/settings/SettingsActivity.cpp
@@ -50,7 +50,6 @@
 #include "activities/util/ConfirmationActivity.h"
 #include "components/UITheme.h"
 #include "fontIds.h"
-#include "util/ChapterTimeEstimate.h"
 #include "util/HeaderDateUtils.h"
 #include "util/ShortcutRegistry.h"
 #include "util/ShortcutUiMetadata.h"
@@ -348,7 +347,7 @@ std::string getSettingValueText(const SettingInfo& setting) {
     const uint8_t value = SETTINGS.*(setting.valuePtr);
     if (setting.nameId == StrId::STR_STATUS_BAR_CHAPTER_PROGRESS) {
       char buf[64];
-      if (ChapterTimeEstimate::formatChapterProgressLabel(value, buf, sizeof(buf))) {
+      if (CrossPointSettings::formatChapterProgressLabel(value, buf, sizeof(buf))) {
         return buf;
       }
     }
